Implements a log mechanism when uploading tsv files
Whenever an exception is thrown, the TSV status on dataset_meta table
is updated with the status code for FAIL (2). Furthermore, the contents
is saved on a table called tsv_logs_table, with fields dataset_id,
row_id and error message. If error is dataset related (did not occur on
any specific row), row_id is set to (-1).

Resolve #43 and #44