Filter: Track/album versions that aren’t allowed
Implements a script to run tests

This script runs a single test against a mock file. We're doing this
outside the test scope in order to catch exceptions, since error
messages returned by the test suite are rather cryptic. When everything
is running fine, a unit test should be performed on the module.

Resolve #79